+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added basic readTheDocs structure for documentation [#12](https://github.com/ie3-institute/powerflow/issues/12)
 
 ### Changed
+- Provide error message when fail calculating the correction [#24](https://github.com/ie3-institute/powerflow/issues/24)
 
 ## [0.4]